Extract routes into an export
This will allow importing in our apps and remove duplication as well as
ensuring that changes sync correctly.
